+++ posthog-node/CHANGELOG.md
@@ -1,3 +1,8 @@
+
+# 4.0.1 - 2024-04-25
+
+1. Prevent double JSON parsing of feature flag payloads, which would convert the payload [1] into 1.
+
 # 4.0.0 - 2024-03-18
 
 ## Added