+++ posthog-react-native/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+# 3.1.1 - 2024-04-25
+
+1. Prevent double JSON parsing of feature flag payloads, which would convert the payload [1] into 1.
+
 # 3.1.0 - 2024-03-27
 
 ## Changed